+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Bug fixes
 * unbind functions before setting property with `set-obj!`
 * make gensyms names shorter (like in CLISP)
+* fix double quasiquote + unquote unquote splice (and other quirks)
 
 ## 0.16.1
 ### Bug fixes